third_party/WebKit/LayoutTests/fast/dom/htmlcollection-non-html-expected.txt

Tests that HTMLCollection only properly contains HTML elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS select.options.length is 0
PASS select.selectedOptions.length is 0
PASS select.options.length is 1
PASS document.images.length is 0
PASS document.images.length is 1
PASS document.forms.length is 0
PASS document.forms.length is 1
PASS document.applets.length is 0
PASS document.applets.length is 1
PASS document.embeds.length is 0
PASS document.embeds.length is 1
PASS document.scripts.length is 3
PASS document.scripts.length is 4
PASS document.links.length is 0
PASS document.links.length is 1
PASS document.links.length is 0
PASS document.links.length is 1
PASS document.anchors.length is 0
PASS document.anchors.length is 1
PASS elem.areas.length is 0
PASS elem.areas.length is 1
PASS elem.rows.length is 0
PASS elem.rows.length is 1
PASS elem.tBodies.length is 0
PASS elem.tBodies.length is 1
PASS elem.cells.length is 0
PASS elem.cells.length is 1
PASS elem.rows.length is 0
PASS elem.rows.length is 1
PASS elem.rows.length is 0
PASS elem.rows.length is 1
PASS elem.rows.length is 0
PASS elem.rows.length is 1
PASS successfullyParsed is true

TEST COMPLETE
